Move FAQ, support, and Discord SSR tests under the 2000-line budget
ssr-render.node.test.ts was still over the node-test ratchet. Those three public-page cases now live in their own file. The snapshot is unchanged.
